Always pass default field's values to __new__()/__init__().
Previously, during Struct instantiation, if values for fields with default values were not explicitly provided, these fields would be initialized but not passed to the Struct's __new__() and __init__() methods. This meant that in order to cover all cases of instantiation, the user had to include default values in the parameter lists for __new__()/__init__().

Under the new behavior, the user should specify all fields as parameters in __new__()/__init__(), but need not include their default values.